refactor(donut): mover DonutCard a Components y actualizar imports
